+++ src/Subscription.ts
@@ -4,7 +4,6 @@ import { Document, Filter } from "mongodb";
 import { ErrorEvents } from "./ErrorEvents";
 import { PromiseTracker } from "./PromiseTracker";
 import { TypedEventEmitter } from "./TypedEventEmitter";
-import { toError } from "./toError";
 
 export interface SubscriptionOptions<TMessage extends Document> {
   /** Local filter to apply on received messages (in-memory). */
@@ -56,7 +55,7 @@ export class Subscription<TMessage extends Document> extends TypedEventEmitter<
     try {
       await this.promises.run(async () => this.callback(message));
     } catch (err) {
-      this.emit("error", toError(err), message);
+      this.emit("error", err as Error, message);
     }
   }
 }