+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Require Node.js >= 18
   - Consumer and subscriber callbacks must return `void`
 - Fix issue where `BatchPublisher` would not publish a message until next `publish` call
+- Remove casting to `Error`
 - Update deps
 
 ## 0.4.0